Move try-except blocks to chech_subject() and get_schoolkid() functions

+++ scripts.py
@@ -14,24 +14,48 @@ class SchoolkidEmptyString(Exception):
     pass
 
 def get_schoolkid(name):
-    if not name:
-        raise SchoolkidEmptyString('Указана пустая строка. Необходимо указать имя ученика.')
-    schoolkid = get_object_or_404(Schoolkid.objects.all(), full_name__contains=name)
-    return schoolkid
+    def check_name(name):
+        if not name:
+            raise SchoolkidEmptyString('Указана пустая строка. Необходимо указать имя ученика.')
+    try:
+        check_name(name)
+    except SchoolkidEmptyString as error:
+        print(error)
+        return False
+    try:
+        schoolkid = get_object_or_404(Schoolkid.objects.all(), full_name__contains=name)
+    except Schoolkid.MultipleObjectsReturned as error:
+        number_of_schoolkids = re.findall(r"\d+", error.args[0])[0]
+        print(f'Найдено {number_of_schoolkids} ученика(ов) с таким именем. Уточните запрос.')
+        return False
+    except django.http.response.Http404:
+        print(f'Ученик с именем {name} не найден. Уточните запрос.')
+        return False
+    else:
+        return schoolkid
 
 
 def check_subject(subject, schoolkid):
     subject = subject.title()
-    subjects = Subject.objects.filter(year_of_study=schoolkid.year_of_study)
-    subjects = [subject.title for subject in subjects]
-    subjects_message = '\n'.join(subjects)
-    if subject not in subjects:
-        raise SubjectDoesNotExist(f'Предмет {subject} не найден в списке изучаемых предметов указанного ученика. Уточните запрос.\n'
-                                  f'Необходимо указать предмет из списка:\n{subjects_message}')
-    return subject
+    def isexist(subject, schoolkid):
+        subjects = Subject.objects.filter(year_of_study=schoolkid.year_of_study)
+        subjects = [subject.title for subject in subjects]
+        subjects_message = '\n'.join(subjects)
+        if subject not in subjects:
+            raise SubjectDoesNotExist(f'Предмет {subject} не найден в списке изучаемых предметов указанного ученика. Уточните запрос.\n'
+                                      f'Необходимо указать предмет из списка:\n{subjects_message}')
+    try:
+        isexist(subject, schoolkid)
+    except SubjectDoesNotExist as error:
+        print(error)
+        return
+    else:
+        return subject
 
 def fix_marks(name):
     schoolkid = get_schoolkid(name)
+    if not schoolkid:
+        return
     marks = schoolkid.mark_set.filter(points__in=[2, 3])
     for mark in marks:
         mark.points = 5
@@ -40,29 +64,18 @@ def fix_marks(name):
 
 def remove_chastisements(name):
     schoolkid = get_schoolkid(name)
+    if not schoolkid:
+        return
     chastisements = schoolkid.chastisement_set.all()
     for chastisement in chastisements:
         chastisement.delete()
 
 
 def create_commendation(name, subject):
-    try:
-        schoolkid = get_schoolkid(name)
-    except SchoolkidEmptyString as error:
-        print(error)
+    schoolkid = get_schoolkid(name)
+    if not schoolkid:
         return
-    except Schoolkid.MultipleObjectsReturned as error:
-        number_of_schoolkids = re.findall(r"\d+", error.args[0])[0]
-        print(f'Найдено {number_of_schoolkids} ученика(ов) с таким именем. Уточните запрос.')
-        return
-    except django.http.response.Http404:
-        print(f'Ученик с именем {name} не найден. Уточните запрос.')
-        return
-    try:
-        subject = check_subject(subject, schoolkid)
-    except SubjectDoesNotExist as error:
-        print(error)
-        return
+    subject = check_subject(subject, schoolkid)
 
     texts = [
         'Молодец!',
